Fixed the bug that you cant move anymore
The problem was that when the new scene load somehow the player input map will disconnect the binds. When you press an input it will trigger but if you want to read the value it is null. I fixed it by forcing the input map to enable when a scene is loaded